refactor curseforge download metrics
